+++ codegen/build_add_groupings_2.py
@@ -88,7 +88,7 @@ def gte(lhs, rhs):
         return True
 
     if isinstance(rhs, int):
-        lhs = int(lhs)
+        lhs = float(lhs)
 
     return lhs >= rhs
 """
